requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
aiosqlite==0.20.0
openai==1.59.3
python-dotenv==1.0.1
pydantic==2.10.4
